DROTH-2706 change to numeric so desimal is possible

+++ digiroad2-oracle/src/main/resources/db/migration/V0_1__create_tables.sql
@@ -563,7 +563,7 @@ CREATE TABLE number_property_value_history
     id          bigint NOT NULL,
     asset_id    bigint NOT NULL,
     property_id bigint NOT NULL,
-    value       bigint,
+    value       numeric,
     grouped_id  bigint DEFAULT 0
 );
